+++ pac4j-oidc/src/main/java/org/pac4j/oidc/federation/entity/EntityConfigurationGenerator.java
@@ -1,0 +1,14 @@
+package org.pac4j.oidc.federation.entity;
+
+/**
+ * Entity configuration generator.
+ *
+ * @author Jerome LELEU
+ * @since 6.4.0
+ */
+public interface EntityConfigurationGenerator {
+
+    String getContentType();
+
+    String generate();
+}